experiment with using installed gtest
Under debian-based systems, the package libgtest-dev can be installed.
It provides the include files, such as gtest.h, in /usr/include/gtest,
and it provides the source files, such as src/gtest-all.cc, in /usr/src/gtest/.

Here we barely start to accomodate this, but it is not urgent.  See
https://github.com/Macaulay2/M2/issues/195

The next step is to check for the presence of src/gtest-all.cc and inform the
configure scripts of mathic and mathicgb where it is.